requirements.txt
backcall==0.1.0
cent==3.0.0
certifi==2018.11.29
chardet==3.0.4
decorator==4.3.0
Django==2.1.4
Faker==0.9.2
gunicorn==19.9.0
HashTable==0.1
idna==2.8
ipython==7.0.1
ipython-genutils==0.2.0
jedi==0.13.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.6
ptyprocess==0.6.0
Pygments==2.2.0
PyJWT==1.7.1
python-dateutil==2.7.3
python-memcached==1.59
pytz==2018.5
requests==2.21.0
simplegeneric==0.8.1
six==1.11.0
text-unidecode==1.2
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
